fixed ects being ints instead of float and made changes to the layout
fixed int intstead of float on ects in course table

changed the template to get some space on the sides so table doesnt take up whole screen.

+++ Course_Finder/schema.sql
@@ -11,7 +11,7 @@ CREATE TABLE Role (
 CREATE TABLE Course (
     id SERIAL PRIMARY KEY,
     duration INT NOT NULL,
-    ECTS INT NOT NULL,
+    ECTS FLOAT NOT NULL,
     name VARCHAR(100) NOT NULL,
     prerequisite_course_id INT, 
     FOREIGN KEY (prerequisite_course_id) REFERENCES Course(id)
